requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-geoip==1.2
python-geoip-geolite2==2015.303
Werkzeug==0.16.0
maxminddb-geolite2
flask-cors